feat: split project into four parts for separate client, game, server, and main
- now using client/game/server DLL for functionality with null implementation if not found
- add client DLL, base implementation, and null client
- add WIP cvar system
- add WIP FilePath class (sandbox to game dir) and FileSystem
- changed global variable formatting to reflect style changes
- adjust some attribute defines to add code section, noinline, noreturn functionality
- add source_location STL replacement for GNU
- removed nodiscard, deprecated, noreturn macros
- remove old game hook
- various other changes here and there

+++ src/core/src/system/system_info.cpp
@@ -114,7 +114,7 @@ void GetCPUInfo(SystemInfo& info)
     } while (current.QuadPart - count.QuadPart < wait.QuadPart);
 
     SetThreadPriority(thread_info, priority);
-    info.cpu_mhz = static_cast<f32>(((__rdtsc() - start) << 5) / 1000000.0);
+    info.cpu_mhz = static_cast<f32>(static_cast<f32>((__rdtsc() - start) << 5) / 1000000.0f);
 }
 
 void GetMemoryInfo(SystemInfo& info)
